Rename conversation_tmp to wizard.
