+++ go/conversation/templates/conversation/conversation_base.html
@@ -14,8 +14,8 @@
         <a href="#">Preview &amp; Activate</a>
     {% else %}
         {# TODO: Make real conversation screens for these? #}
-        <a href="{% url 'conversations_tmp:message' conversation.key %}">Messages</a>
-        <a href="{% url 'conversations_tmp:contacts' conversation.key %}">Contacts</a>
+        <a href="{% url 'wizard:message' conversation.key %}">Messages</a>
+        <a href="{% url 'wizard:contacts' conversation.key %}">Contacts</a>
         <a href="{% conversation_screen conversation %}">Preview &amp; Activate</a>
     {% endif %}
 {% endblock %}
