+++ go/conversation/templates/conversation/dashboard.html
@@ -11,7 +11,7 @@
 
 
 {% block content_actions_left %}
-    <a href="{% url 'conversations_tmp:details' %}" class="btn btn-primary">New conversation</a>
+    <a href="{% url 'wizard:details' %}" class="btn btn-primary">New conversation</a>
 {% endblock %}
 
 {% block content_actions_right %}
