+++ go/templates/app.html
@@ -23,7 +23,7 @@
                 <b class="caret"></b>
             </a>
             <ul class="dropdown-menu" role="menu">
-                <li><a href="{% url 'conversations_tmp:pricing' %}">Pricing</a></li>
+                <li><a href="{% url 'wizard:pricing' %}">Pricing</a></li>
             </ul>
         </li>
