+++ go/urls.py
@@ -28,8 +28,7 @@ urlpatterns = patterns(
     url(r'^conversations/',
         include('go.conversation.urls', namespace='conversations')),
     url(r'^channels/', include('go.channel.urls', namespace='channels')),
-    url(r'^conversation-tmp/', include('go.conversation_tmp.urls',
-        namespace='conversations_tmp')),
+    url(r'^wizard/', include('go.wizard.urls', namespace='wizard')),
     url(r'^contacts/', include('go.contacts.urls', namespace='contacts')),
     url(r'^account/', include('go.account.urls', namespace='account')),
     url(r'^accounts/', include('registration.backends.default.urls')),
