+++ go/wizard/templates/wizard_views/wizard_1_details.html
