+++ go/wizard/templates/wizard_views/wizard_2_message.html
