+++ go/wizard/templates/wizard_views/wizard_base.html
@@ -13,8 +13,8 @@
         <a href="#">Contacts</a>
         <a href="#">Preview &amp; Activate</a>
     {% else %}
-        <a href="{% url 'conversations_tmp:message' campaign_key %}">Messages</a>
-        <a href="{% url 'conversations_tmp:contacts' campaign_key %}">Contacts</a>
+        <a href="{% url 'wizard:message' campaign_key %}">Messages</a>
+        <a href="{% url 'wizard:contacts' campaign_key %}">Contacts</a>
         <a href="{% conversation_screen conversation %}">Preview &amp; Activate</a>
     {% endif %}
 {% endblock %}
